pyproject.toml
[tool.poetry]
name = "test-auth-service"
version = "0.1.0"
description = ""
authors = ["bm666 <artem.vostrikov.777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
alembic = {extras = ["async"], version = "^1.15.2"}
asyncpg = "^0.30.0"
python-jose = "^3.4.0"
httpx = "^0.28.1"
uvicorn = "^0.34.2"
pydantic = {extras = ["email"], version = "^2.11.4"}
pydantic-settings = "^2.9.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aioredis = "^2.0.1"
redis = {extras = ["async"], version = "^6.0.0"}
pytest-asyncio = "^0.26.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88